Move redundant code to base class defs
